Completed the image functionality on the upload component
