+++ app/components/Upload.jsx
@@ -1,5 +1,5 @@
 "use client"
-import React, { useState, useEffect, useRef } from "react";
+import  { useState, useEffect, useRef } from "react";
 import Image from "next/image";
 import Cropper from "react-easy-crop";
 import * as React from "react";
@@ -8,28 +8,109 @@ import Slider from "@mui/material/Slider";
 import Modal from "@mui/material/Modal";
 const Upload = (props) => {
   const inputRef = useRef();
-    
+  const [image, setImage] = useState(null);
+  const [croppedImage, setCroppedImage] = useState(null);
+  const [croppedArea, setCroppedArea] = useState(null);
+  const [crop, setCrop] = useState({ x: 0, y: 0 });
+  const [zoom, setZoom] = useState(1);
+  const [rotation, setRotation] = useState(0);
+  const [open, setOpen] = useState(false);
 
-  const handleImageChange = (e) => {
-    const file = e.target.files[0];
-    if (file) {
-      const imageUrl = URL.createObjectURL(file);
-      props.setSelectedImage(imageUrl);
-      // Reset the input value to null to allow the same file to be selected again
-      if (inputRef.current) {
-        inputRef.current.value = null;
-      }
-    }
-  };
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
+
+  // const handleImageChange = (e) => {
+  //   const file = e.target.files[0];
+  //   if (file) {
+  //     const imageUrl = URL.createObjectURL(file);
+  //     props.setSelectedImage(imageUrl);
+  //     // Reset the input value to null to allow the same file to be selected again
+  //     if (inputRef.current) {
+  //       inputRef.current.value = null;
+  //     }
+  //   }
+  // };
 
   const selectedFilePopup = () => {
     inputRef.current.click();
   };
 
+    const onCropComplete = (croppedAreaPercentage, croppedAreaPixels) => {
+    console.log("pixels", croppedAreaPixels);
+    setCroppedArea(croppedAreaPixels);
+  };
+
+  const onSelectFile = (event) => {
+    const file = event.target.files;
+    if (file && file.length > 0) {
+      const reader = new FileReader();
+      reader.readAsDataURL(file[0]);
+      reader.addEventListener("load", () => {
+        setImage(reader.result);
+        handleOpen(); // opening the modal when user select a file
+      });
+    }
+  };
+
+  const getCroppedImg = async () => {
+    if (!image || !croppedArea) return;
+
+    const imageElement = document.createElement("img");
+    imageElement.src = image;
+
+    const canvas = document.createElement("canvas");
+    const ctx = canvas.getContext("2d");
+
+    const scaleX = imageElement.naturalWidth / imageElement.width;
+    const scaleY = imageElement.naturalHeight / imageElement.height;
+
+    canvas.width = croppedArea.width;
+    canvas.height = croppedArea.height;
+
+    ctx.save();
+    ctx.translate(croppedArea.width / 2, croppedArea.height / 2);
+    ctx.rotate((rotation * Math.PI) / 180);
+    ctx.translate(-croppedArea.width / 2, -croppedArea.height / 2);
+    ctx.drawImage(
+      imageElement,
+      croppedArea.x * scaleX,
+      croppedArea.y * scaleY,
+      croppedArea.width * scaleX,
+      croppedArea.height * scaleY,
+      0,
+      0,
+      croppedArea.width,
+      croppedArea.height
+    );
+    ctx.restore();
+
+    return new Promise((resolve, reject) => {
+      canvas.toBlob((blob) => {
+        if (!blob) {
+          reject(new Error("Canvas is empty"));
+          return;
+        }
+        const croppedUrl = URL.createObjectURL(blob);
+        resolve(croppedUrl);
+      }, "image/jpeg");
+    });
+  };
+
+  const handleShowCroppedImage = async () => {
+    const croppedImageUrl = await getCroppedImg();
+    setCroppedImage(croppedImageUrl);
+    props.setSelectedImage(croppedImageUrl);
+    handleClose(); // closing the modal when user select a file
+  };
+
+  const rotateLeft = () => setRotation((prev) => (prev - 90) % 360);
+
+
+
 
   return (
     <div>
-      <div className="flex flex-col bg-transparent border py-10 mt-2 rounded-md">
+      <div className="flex flex-col bg-transparent border py-10 mt-2 rounded-md hover:border-blue-500 h-full w-full ">
         <label className="flex flex-col justify-center gap-1 items-center bg-transparent p-4 text-2xl text-gray-600 cursor-pointer whitespace-nowrap ">
           {props.selectedImage ? (
             <>
@@ -37,15 +118,15 @@ const Upload = (props) => {
                 type="file"
                 className="hidden"
                 ref={inputRef}
-                onChange={handleImageChange}
+                onChange={onSelectFile}
               />
-              <button onClick={selectedFilePopup}>
+              <button onClick={selectedFilePopup}  className="h-full w-full">
               <Image
-                className="w-full h-auto  object-cover"
+                className="object-cover rounded-md w-full h-full "
                 src={props.selectedImage}
                 alt="p"
-                width={150}
-                height={150}
+                width={500}  // Adjust these values as needed
+                height={500} // Adjust these values as needed
               />
               </button>
             </>
@@ -54,7 +135,7 @@ const Upload = (props) => {
               <input
                 type="file"
                 className="hidden"
-                onChange={handleImageChange}
+                onChange={onSelectFile}
               />
               <svg
                 xmlns="http://www.w3.org/2000/svg"
@@ -80,6 +161,58 @@ const Upload = (props) => {
           )}
         </label>
       </div>
+       <Modal open={open} onClose={handleClose}>
+        <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 bg-white p-5 rounded-md shadow-lg">
+          {image && (
+            <div className="flex flex-col items-center">
+              <div className="w-80 h-80 overflow-hidden relative">
+                <Cropper
+                  image={image}
+                  crop={crop}
+                  zoom={zoom}
+                  rotation={rotation}
+                  aspect={1}
+                  onCropChange={setCrop}
+                  onZoomChange={setZoom}
+                  onCropComplete={onCropComplete}
+                  className="w-full h-full"
+                />
+              </div>
+              <Box sx={{ width: 300, mt: 2 }}>
+                <Slider
+                  value={zoom}
+                  min={1}
+                  max={3}
+                  step={0.1}
+                  onChange={(e, zoom) => setZoom(zoom)}
+                  color="secondary"
+                />
+              </Box>
+              <div className="flex mt-2 space-x-2">
+                <button
+                  className="px-4 py-2 bg-blue-500 text-white rounded"
+                  onClick={rotateLeft}
+                >
+                  Rotate
+                </button>
+
+                <button
+                  className="px-4 py-2 bg-blue-500 text-white rounded"
+                  onClick={selectedFilePopup}
+                >
+                  New
+                </button>
+              </div>
+              <button
+                className="px-4 py-2 mt-3 bg-blue-500 rounded-md text-white hover:bg-blue-700"
+                onClick={handleShowCroppedImage}
+              >
+                Conferma
+              </button>
+            </div>
+          )}
+        </div>
+      </Modal>
     </div>
   );
 };
